refactor(ListBoxWithTags): UXD-927 extract renderTrigger to separate file
